moved old Expectation in CoreExpectation.php and made Expectation.php a decorator for it
