+++ src/Factories/TestCaseFactory.php
@@ -144,8 +144,8 @@ final class TestCaseFactory
          * @return mixed
          */
         $test = function () use ($chains, $proxies, $factoryTest) {
-            $chains->chain($this);
             $proxies->proxy($this);
+            $chains->chain($this);
 
             /* @phpstan-ignore-next-line */
             return call_user_func(Closure::bind($factoryTest, $this, get_class($this)), ...func_get_args());
